Moved read function to local only builtins
